Add files via upload
Job: Implementing a client to send requests to the Fair Sign API using the client's HTTP and JSON serialization libraries. The client provides thread-safe execution and support for limiting the number of API requests. A method has been implemented to create a document for the introduction into circulation of goods produced in the Russian Federation. The code is written using an object-oriented approach and is as convenient as possible for further expansion of the functionality.

Code functionality: Client implementation for interacting with the Fair Sign API using the HTTP client and JSON serialization libraries. The client provides a method for creating a document for the entry into circulation of goods produced in the Russian Federation, and provides thread-safe execution and support for limiting the number of API requests.